add item decoration in MainWalletActivity

+++ iconex/src/main/java/foundation/icon/iconex/util/ScreenUnit.java
@@ -4,7 +4,7 @@ import android.content.Context;
 import android.util.TypedValue;
 
 public class ScreenUnit {
-    public static int dp2px (Context context, int dp) {
+    public static int dp2px (Context context, float dp) {
         return (int) TypedValue.applyDimension(
                 TypedValue.COMPLEX_UNIT_DIP, dp,
                 context.getResources().getDisplayMetrics());
